LOOKDEVX-1847 - Fix rendering of gltf texture nodes
We previously were adding default geompropvalue nodes on dangling texcoord inputs in the source MaterialX document.

We now let them be, and process them post shadergen as re-implementation of the
texcoord node into geompropvalue code. This allows dealing with the
texcoord nodes embedded in the glTf texture nodes added to MaterialX
1.38.6.