EMSUSD-604 fix crash when duplicating a proxy shape
The crash occurred in deeply nested callbacks-within-callbacks. There were
attributes being set which triggered dirtynes, which triggered computation,
which triggered atribute updates, which triggered proxy accessor, etc. In the
end, some code accessed the proxy shape Maya node while it was not in a valid
state to be accessed.

The ultimate cause was that, surprinsingly, never during the duplication of a
node does the node has its final name, in any of the usual callbacks. During
the postConstruct callback, it doesn't have a name. During the copyInternal
callback, it does not have a name. But the proxy shape wants to know its
ownname when it registers some callbacks. The end-result was that during save,
the node was discovering that it had a new name and it brought the house down.

The fix is to be notified about node naming and renaming as soon as they
happen. As such, a callback when renamed is added that fixes the issue.

Also, discovered that the way the stage were held during saving was insecure
because they used weak pointer that could become null. Use a referencing
pointer instead to avoid losing stages mid-save.

Add a unit test for the crash condition.