EMSUSD-647 fix copying the proxy shape node
When Maya copies a Maya node, it actually export the selected nodes into a temporary file. When the node is a MayaUSD node, this export automatically prompt the user to save the dirty USD files. To avoid this, we now detect that we are exporting due to a copy operation and force saving the USD layer into the Maya scene file.