EMSUSD-848 fix edit target after layer clear
When clearing a layer that had sub-layers and one of those sub-layer was the edit target, we would no longer have a valid edit target. Fix this by putting the target on a valid layer, for now the root (since it cannot be muted and thus is valid). Also, remember the edit target so that undo can restore it.

A the same time, split the functions in the command class to have each function do only one thing.