EMSUSD-671 fix reloading scene with layers saved in the Maya scene
When reloading a scene that contained layers serialized directly in the scene,
sub-layers would be lost because they were anonymous layers but on reload that
status was lost and the sub-layer would be lost whenever the stage got recomputed.

Fix this by remembering that the sub-layer are anonymous on reload.

Add unit test to verify lost layer fix

- Add comments to the layer manager serialization test and fix the layer being targeted.
- Refactor common code in proxy shape saving tests.
- Add more checks in the proxy shape saving test to verify that the contents are correct afetr reload and after recompute.