EMSUSD-65 wait cursor during long VP2 updates
When a command may trigger a long viewport update, set a flag so that the proxy render delegate shows the wait cursor while it is updating the viewport data. We make the assumption that long-duration commands will generally cause long-duration viewport updates. This is done via the already-existing LongDurationCommand base class and wait cursor functions.

- Modify the WaitCursor helper class to take a flag to control if the cursor is shown or not.
- This allows conditionallly showing the wait cursor based on a condition.
- Add a function to tell the render delegate that it should expect its update to take a long time.
- Make the wait cursor function used by long-duration commands tell the render delegate to expect a long update.