MAYA-132438 fix slow scene load performance
Maya scene load without any USD data is slower when the MayaUSD plugin is loaded.

The reason behind the bad performance is that the UsdStageMap class constantly rebuilds itself because it thinks it is out of date. The reason is that its UsdStageMap::proxyShape() function assumes that it is always called to retrieve an existing proxy shape. So when it fails to find the requested proxy shape, it assumes its mapping between UFE paths and DAG paths must be stale. Those assumptions are false.

The culprit is the following sequence:
- Adding Maya nodes cause a UFE AddObject notification to be sent.
- The Maya outliner UI wants to know if this object might be a UFE gateway node.
- It asks UFE if is a gateway node.
- This ends-up being routed to MayaUSD.
- ProxyShapeSceneSegmentHandler::isGateway_() calls getStage(path) with the UFE path.
- This ends-up calling UsdStageMap::proxyShape().
- Which triggers the cache rebuilds, as explained above.

While we need to revisit how the UsdStageMap works, this would be a somewhat big and risky endeavour at this point in time. The simple fix is to not rebuild the cache when we know there are no proxy shape in existence. This fixes the performance problem for all Maya scene without USD data.