EMSUSD-988 Maya node observer classes
Add new classes to observe Maya nodes of a given type or a given Maya node. The
observer forward changes to registered listeners. The goal is to standardize how
a node is observed and to share the notification between multiple listeners.

This will be used by multiple classes in a subsequent pull request:
MayaUsdProxyShapeBase, UsdStageMap and MayaSessionState.

The MayaNodeTypeObserver allows listening to new nodes being created or
removed. The MayaNodeObserver allows listening to the node being renamed or
reparented. Many use cases need to track the DAG path, UFE path or USD path of
objects, mostly the USD stage itself.