EMSUSD-998 better universal manip undo redo
Fix the complicated handling of undo and redo vs the universal manipulator. The
code failed to do the right thing when half of the attributes are already
existed or when more complex composite manipulations were done, for example
when scaling with the universal manipulator.

We make the transformation commands more robust vs the ordering of calls to
support all use cases. They still handle the bug related to object-space
translation redo: when redoing an object-space translation, Maya passes in
incorrect values to the command that must be ignored.

Apply fix to all transform commands. This way they all use the same pattern
from UsdSetXformOpUndoableCommandBase to manage any order of calls to
execute/set/undo/redo: use in the CommonAPI commands, matrix commands
and Maya tranform-stack commands.