EMSUSD-1017 Fix reverting multiple files
When reverting multiple USD layers, reverting the first layer would trigger the UI to update, thus making the the element holding the other layers invalid before they were processed.

To avoid this, add a mechanism to delay the execution of commands in the command-hook so that all item can be processed before the first command is executed.

In the Maya implementation of the command hook, the MEL and Python commands are accumulated when the commands are delayed and executed once the delayed commands are triggered.