Restrict UFE nodes according to proxyNode's primPath
Only create UFE nodes that are ancestors or descendants of the proxy
node's primPath. This addresses a performance issue we encountered in
our Reference Assembly workflows. For these workflows, there were
reference assemblies that were referencing very large Usd scenes, but
drawing a very small subset of it. While Maya would only image the
subset, UFE was populating every node from the Usd stage.

This caused several performance problems, including:
* viewport -> panel -> perspective slowness due to it traversing all the
UFE nodes looking for cameras
* selection with the transform tool

This fix limits creation of UFE nodes to those "relevant" to the prim
path

This fix also changes behavior of imaging a proxy node with an invalid
prim path. Previously, a non-empty invalid path would cause the entire
stage to be imaged. With this change, we image nothing.