Author references but do not have the UsdStage compose
The stage that maya uses when exporting to use has load policy set to `UsdStage::LoadAll`.  Whenever we would author a reference to a model, that would result in the referenced model being loaded and then composed into the stage.  This meant that at export time was now dependent on the models we are referencing, rather than solely dependent on the contents of the current maya file.

With this change, we call `Unload` on the referencing prim before authoring the reference.  This sets the load policy to not load the prim when we eventually do author the reference.