EMSUSD01118 instance selection highlight
Detect the need to redo the selection highlight by detecting if the instance count changed. Unfortunately, the instancer index version only changes when an instance was toggle off, not when togged on. Might be a bug or not in USD, but we work around it by tracking the instance count too.