EMSUSD-1118 fix affect unit tests
Explicitly set the selection instead of trying to change the images because the behavior is different between Maya 2022 and the later versions because the fix is only available starting with Maya 2023. It is simpler to make the text behave the same in all version by explicitly selecting the object that the image expect to be selected. (Even if it was by mistake in Maya 2022...)