EMSUSD-1258 extend the delete cmd edit routing
The new code for doing edit routing using an edit target directly was
not supporting the usual use-case of providing the target by the layer
identifier or layer handle.

It also always allowed stronger opinions, even though we only need it if
the target is within a reference or payload.

This fixes both issues: delete can now be routed like other commands by
ID or layer handle and will only allow stronger opinions when targeting
a non-local layer.

Extending using an edit target directly in all other commands will be
done separatey in the future.