EMSUSD-1257 Fast Python routing
Add a C++ edit router that unconditionally route a command to a layer
for a given stage. Expose this router to Python.

This allows a fast edit router even in Python. We avoid creating USD and
Python dictionaries and just use the given layer for the command. This
ensures that the routing is very efficient.

Update the docs to describe this.