EMSUSD-1260 fix rotation when prim already has single-axis rotation
Force the prim to have a three-axis rotation when creating a rotation
command.

The fallback transform implementation derives from the Maya one, but
wants the old behavior of not changing how the ops work. So we add a
virtual function to detect this.

Add a unit test that test the situation.