Add remapUVSetsTo option to UsdImport
Our users have requested that UV Sets that were renamed on export round
trip cleanly on import. The Maya:name metadata accomplishes this, but
only for assets that were exported after the introduction of this code.
This allows us to specify a remapping for UV Set names on import so
older assets work.