EMSUSD-1160 improve Maya pivot support
Multiple tools require the use of Maya-style pivots. The difference
between Maya-style and USD-style is that Maya has two separate pivots
for rotation and scaling while USD has a single pivot for both. In
addition, Maya has pivot-compensation attributes.

When there is already a USD-style pivot authored, we need to convert it
to the Maya-style pivots when these tools try to modify the object. We
detect that such a conversion is needed and adjust the values when they
are read and do the conversion the first time a tools tries to write the
values.

Fix center pivot test to take into account the new behaviour.
Add new unit test for pivot conversion.