EMSUSD-1357 material scope when no default prim given
Change when the default prim is selected when the export job has no
default prim explicitly given. It is now chosen at the beginning
the export. This change allows the material scope to know that it must
be placed under that default prim.

Take into account root prim and export roots. When the root prim or
export roots are explicitly given, use them as the default prim if no
explicit default prim was given.

Fix all unit tests affected by the change to where the material scope is
placed. Some tests are fixed by explicitly requested that no default
prim be used, some are fixed by changing where the material scope is
checked to be. These two variations allow testing the new behaviour
without having to write new unit tests.