EMSUSD-1467 fix multi-layer parenting
The parenting code was not copying the data correctly. Changed to use
the similar code as in the duplicate command. The duplicate command
could not be used because the parenting command targets an arbitrary
destination.

Added a unit test that exercises the case where a prim is in multiple
layers and targeting the session layer. The old code would lose data.

Nested edit router failed to route to the correct layer and woudl result
instead in an infinite loop.