Remove matrix multiply operator overloads that potentially lose precision
The overloads of the methods that take and return a GfVec3f can cause
inadvertent loss of precision. This change updates call sites where the
result of a call to one of these overloads now needs to be converted to
GfVec3f. The advantage of requiring these conversions is that the loss
of precision is explicit.