EMSUSD-1477 merge job context arguments with user arguments
The VRay plugin provides the chaser name by the user arguments instead
of returning it in the plug-in job-context callback. (How they inject
the chaser name in the arguments is still a mystery.)

The code used to overwrite such user arguments with the job-context
arguments. (By calling VtDictionaryOver.) Now we instead merge the
dictionaries with code refactored from the code used to merge job
context arguments.

Do the same for the import context.