EMSUSD-1619 fix removal on multiple layers
When removing multiple layers the removal must not be delayed. This is due to the low-level implementation of layer removal: it uses the parent layer and the index of the sub-layer to be removed instead of its name. Wehn commands are delayed, the index would become incorrect. By not delaying command, the index is properly calculated.