EMSUSD-1735 Correct matrix manipulation
When a USD prim is transformed using a matrix, correctly read the initial matrix values when teh transform is time-sampled.

Note that we still write the new matrix in the USD default value *not* at the time sample. So it only "works" like it used to "works" in the sense that when translating, for example, the other elements of the matrix are not lost, but the new value is not time-sampled.

- Revert the change in the matrix implementation of the UFE transform3d interface to read the initial matrix value at the read-time not the write-time.
- Currently, the write time is always the USD default-time.
- Add a unit test to cover this case.