EMSUSD-1570 add a class-prims filter to USD UFE
- Add a getUsdPredicate helper function to UsdUfe to create the USD
  predicate corresponding to a list of UFE filters.
- Add the "Class Prims" filter to the list of USD filters in the
  hierarchy handler.
- Use the helper to create the USD predicate.
- Rely on the USD predicate having a correct predicate for inactive prim
  instead of using an explicit flag.
- Adjust the context menu for class prims.
- Add a unit test for the class prims filter.

Work around the problem that Maya does not correctly handle UFE filters.
Maya only handles one UFE filter. If there are more than one filters,
only one is handled.

To work around this issue, we manually ask the outliner what are the
desired settings and use these settings as the default values for the
filters. When the outliner fails to handle multiple filters, it will
end-up using the default values, which will be the desired values.

- Add a CMake compiler definition on the project.
- The script to update the filter need to be protected from infinite
  recursion, no need to run it on idle.
- Running it on idle would be making the filter new value not always
  being applied at the proper time. Easily visible when reload a scene.