Fix AE extra attribute section.
A previous change broke which attributes are added to the extra attributes section. That is because the default function to create an attribute UI returns None and was interpreted as no UI being created.

This is a unfortunate consequence of the design: the return value is interpreted as sucsess or failure of creating a UI and then, on success, custom callbacks are connected to that UI. But for the default fallback, those callbacks must not be added, so it returns None, which was interpreted as failure. Given that it is the final fallback, that had no consequences, but a code change made registering the attribute as being handled was no longer done properly.

Restore the old code, which was adding the attribute to teh list of known, handled attributes. Also, don't add supressed attributes in the extra attribute section.