Fix UsdMayaPrimWriter and UsdMayaPrimReader Python Wrappings
The PrimReaderWrapper<T> and PrimWriterWrapper<T> helper classes ignored the template parameter when deriving from TfPyPolymorphic. This caused unexpected duplicate entries in boost::python's type conversion table. boost::python silently accepted this, but the new pxr_boost::python in OpenUSD 24.11 flagged this with an assert in debug mode.

The wrappings for ShaderReaderWrapper and ShaderWriterWrapper incorrectly specified PrimReaderWrapper<> and PrimWriterWrapper<> as their bases. For example, this meant that passing an instance of a ShaderReader to a a Python-wrapped C++ function that takes a UsdMayaPrimReader would fail. This doesn't show up in practice since clients aren't expected to create their own instances of these types but it seemed good to fix anyway.