EMSUSD-2160 select conflicted expression members
When there is both include or exclude members and an expression in a collection, there is no USD API to retrieve only the members that are chosen bythe expression.

To work aroud this, we create a tempoary prim with a temporary collection with only the expression and ask that temporary collection for its membership.