EMSUSD-1896 avoid crash when shader has no uvLink
The uvLink command can return nothing if it cannot determine the UV link
of the shader node. This would break assumptions in the export code and
lead to a crash when accessing missing data. Handle the case where the
uvLink command returns nothing to prevent that crash.

Add a unit test that reproduced the crash and now no longer crashes.