Add support for Accessibility Export
We recently added support for Accessibility metadata and are trying to enable the creation of more accessible USD files out of Maya.

This first implementation supports adding data on the default prim and corresponds with a matching PR we will be making to Blender. This adds two new fields to the Advanced group.

Our goal here is to enable third party developers and internal pipelines to store the data in their USD files so they can surface it to accessibility runtimes.

Since Accessibility was only added in USD 25.5, I currently implemented this by ad-hoc writing the attributes. I realize that Maya 2026.2 has 25.5 mode, but many of our third party developers aren't yet able to upgrade to Maya 2026 or run in the OpenUSD 25.5 mode. Once Maya has 25.5 ubiquitously available, I can replace this with the actual AccessibilityAPI calls, but in the meantime the test validates the result against USD 25.5 for your CI.