EMSUSD-2969 bake export unit
When exporting with the -unit or -upAxis flags, bake the scaling and
rotation into the vertices positions instead of having a global rotation
or scaling on the root nodes. That is the preferred way user want to
work.

In particular, adding transformations (scaling or rotation) on the root
nodes clashes with other USD layers that studio want to put on top.
Those layer may come from other sources. So baking the position is
preferred as it does not affect other layers that are composed over the
exported data.